# api/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SOL ⇄ DCC Bridge API",
    "description": "REST API for the SOL-Gateway-DCC cross-chain bridge. Provides deposit/redeem instruction generation, transfer tracking, fee estimation, health monitoring, and admin operations.\n\n**Security Note:** This API never holds funds. All instructions are generated for client-side signing.\n\n### Rate Limits\n- **Global:** 100 requests / 15 minutes per IP\n- **Deposit/Redeem:** 10 requests / minute per IP\n- Internal (localhost/Docker) requests bypass rate limiting.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "SOL-DCC Bridge Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    },
    {
      "url": "https://bridge-api.example.com/api/v1",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/redeem": {
      "post": {
        "tags": ["Redeem"],
        "summary": "Generate burn/redeem instruction",
        "description": "Creates a DCC burn instruction. The user signs this on the DCC side to burn wSOL.DCC, which triggers validators to unlock SOL on Solana.",
        "operationId": "createRedeem",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RedeemRequest" },
              "example": {
                "sender": "3P8eZWKG5BGQM7VuvCFtK6GbHFVR3z5DTb",
                "solRecipient": "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU",
                "amount": 1.0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Redeem instruction generated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RedeemResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/BridgePaused" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/cr-stable/mint": {
      "post": {
        "tags": ["CR Stable"],
        "summary": "Generate CR Stable mint instruction",
        "description": "Creates a deposit instruction for minting CR Stable (CRS) tokens. Accepts USDC or USDT as the source token — SOL must be swapped to USDC/USDT client-side (via Jupiter) before calling this endpoint.\n\nReturns PDA addresses and fee breakdown for client-side transaction construction.",
        "operationId": "mintCRStable",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CRStableMintRequest" },
              "example": {
                "sender": "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU",
                "recipientDcc": "3P8eZWKG5BGQM7VuvCFtK6GbHFVR3z5DTb",
                "amount": 100,
                "splMint": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mint instruction generated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CRStableMintResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/cr-stable/info": {
      "get": {
        "tags": ["CR Stable"],
        "summary": "Get CR Stable token info",
        "description": "Returns CR Stable token metadata, backing mechanism details, current reserve status, and fee schedule.",
        "operationId": "getCRStableInfo",
        "responses": {
          "200": {
            "description": "CR Stable token info",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CRStableInfoResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/dusd/mint": {
      "post": {
        "tags": ["Decentral USD"],
        "summary": "Generate DUSD mint instruction",
        "description": "Creates a deposit instruction for minting Decentral USD (DUSD) tokens. Accepts USDC or USDT as the source token (1:1 backing). Returns PDA addresses and fee breakdown for client-side transaction construction.",
        "operationId": "mintDUSD",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DUSDMintRequest" },
              "example": {
                "sender": "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU",
                "recipientDcc": "3P8eZWKG5BGQM7VuvCFtK6GbHFVR3z5DTb",
                "amount": 100,
                "splMint": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mint instruction generated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DUSDMintResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/dusd/redeem": {
      "post": {
        "tags": ["Decentral USD"],
        "summary": "Generate DUSD redeem instruction",
        "description": "Burns DUSD tokens and unlocks the underlying USDC or USDT collateral on Solana. Returns fee breakdown and output details.",
        "operationId": "redeemDUSD",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DUSDRedeemRequest" },
              "example": {
                "sender": "3P8eZWKG5BGQM7VuvCFtK6GbHFVR3z5DTb",
                "recipientSol": "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU",
                "amount": 100,
                "outputMint": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Redeem instruction generated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DUSDRedeemResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/dusd/info": {
      "get": {
        "tags": ["Decentral USD"],
        "summary": "Get Decentral USD token info",
        "description": "Returns DUSD token metadata, backing mechanism details, current reserve status, and fee schedule.",
        "operationId": "getDUSDInfo",
        "responses": {
          "200": {
            "description": "DUSD token info",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DUSDInfoResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AdminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key",
        "description": "Admin API key. Can also be passed as a Bearer token in the Authorization header."
      }
    },
    "schemas": {
...
      "RedeemRequest": {
        "type": "object",
        "required": ["sender", "solRecipient", "amount"],
        "properties": {
          "sender": {
            "type": "string",
            "minLength": 20,
            "maxLength": 64,
            "description": "DCC wallet address (sender of wSOL.DCC)"
          },
          "solRecipient": {
            "type": "string",
            "minLength": 32,
            "maxLength": 44,
            "description": "Solana wallet address to receive unlocked SOL"
          },
          "amount": {
            "type": "number",
            "minimum": 0.001,
            "maximum": 1000,
            "description": "Amount in wSOL.DCC to burn"
          }
        }
      },
      "RedeemResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "instruction": {
            "type": "object",
            "description": "DCC burn instruction for client-side signing"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "bridgeContract": { "type": "string", "description": "DCC bridge contract address" },
              "wsolAssetId": { "type": "string", "description": "wSOL.DCC asset ID on DCC" },
              "amountUnits": { "type": "integer", "description": "Amount in smallest units" },
              "solRecipient": { "type": "string" },
              "estimatedUnlockTime": { "type": "string", "example": "3-10 minutes" },
              "estimatedFee": { "type": "string", "example": "0.005 DCC" }
            }
          }
        }
      },
...
      "CRStableMintRequest": {
        "type": "object",
        "required": ["sender", "recipientDcc", "amount", "splMint"],
        "properties": {
          "sender": {
            "type": "string",
            "minLength": 32,
            "maxLength": 44,
            "description": "Solana wallet address (base58 public key)"
          },
          "recipientDcc": {
            "type": "string",
            "minLength": 20,
            "maxLength": 64,
            "description": "DCC wallet address to receive CRS tokens"
          },
          "amount": {
            "type": "number",
            "minimum": 1,
            "description": "Amount in USD value (e.g. 100 for $100 USDC)"
          },
          "splMint": {
            "type": "string",
            "enum": [
              "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
              "Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB"
            ],
            "description": "SPL token mint address. Only USDC and USDT are accepted."
          }
        }
      },
      "CRStableMintResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "metadata": {
            "type": "object",
            "properties": {
              "programId": { "type": "string", "description": "Bridge program ID" },
              "bridgeConfig": { "type": "string", "description": "BridgeConfig PDA address" },
              "vault": { "type": "string", "description": "Vault PDA address" },
              "userState": { "type": "string", "description": "UserState PDA address" },
              "amountLamports": { "type": "string", "description": "Amount in smallest token units (6 decimals)" },
              "splMint": { "type": "string", "description": "Source SPL token mint address" }
            }
          },
          "crStable": {
            "type": "object",
            "properties": {
              "mintAmount": { "type": "string", "description": "CRS tokens to be minted (after fee)", "example": "99.900000" },
              "feeAmount": { "type": "string", "description": "Fee deducted", "example": "0.100000" },
              "feeRate": { "type": "number", "description": "Fee rate as decimal", "example": 0.001 },
              "symbol": { "type": "string", "example": "CRS" }
            }
          }
        }
      },
      "CRStableInfoResponse": {
        "type": "object",
        "properties": {
          "symbol": { "type": "string", "example": "CRS" },
          "name": { "type": "string", "example": "CR Stable" },
          "fullName": { "type": "string", "example": "Stable CR Coin" },
          "decimals": { "type": "integer", "example": 6 },
          "description": { "type": "string" },
          "backing": {
            "type": "object",
            "properties": {
              "acceptedTokens": { "type": "array", "items": { "type": "string" }, "example": ["USDC", "USDT"] },
              "mechanism": { "type": "string", "example": "1:1 collateralized" },
              "vaultAddress": { "type": "string", "description": "Solana vault address holding reserves" }
            }
          },
          "fees": {
            "type": "object",
            "properties": {
              "mintFee": { "type": "string", "example": "0.10%" },
              "redeemFee": { "type": "string", "example": "0.10%" },
              "minimumMint": { "type": "number", "example": 1 }
            }
          }
        }
      },
      "DUSDMintRequest": {
        "type": "object",
        "required": ["sender", "recipientDcc", "amount", "splMint"],
        "properties": {
          "sender": {
            "type": "string",
            "minLength": 32,
            "maxLength": 44,
            "description": "Solana wallet address (base58 public key)"
          },
          "recipientDcc": {
            "type": "string",
            "minLength": 20,
            "maxLength": 64,
            "description": "DCC wallet address to receive DUSD tokens"
          },
          "amount": {
            "type": "number",
            "minimum": 1,
            "description": "Amount in USD value (e.g. 100 for $100 USDC)"
          },
          "splMint": {
            "type": "string",
            "enum": [
              "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
              "Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB"
            ],
            "description": "SPL token mint address. Only USDC and USDT are accepted."
          }
        }
      },
      "DUSDMintResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "metadata": {
            "type": "object",
            "properties": {
              "programId": { "type": "string", "description": "Bridge program ID" },
              "bridgeConfig": { "type": "string", "description": "BridgeConfig PDA address" },
              "vault": { "type": "string", "description": "Vault PDA address" },
              "userState": { "type": "string", "description": "UserState PDA address" },
              "amountLamports": { "type": "string", "description": "Amount in smallest token units (6 decimals)" },
              "splMint": { "type": "string", "description": "Source SPL token mint address" }
            }
          },
          "dusd": {
            "type": "object",
            "properties": {
              "mintAmount": { "type": "string", "description": "DUSD tokens to be minted (after fee)", "example": "99.900000" },
              "feeAmount": { "type": "string", "description": "Fee deducted", "example": "0.100000" },
              "feeRate": { "type": "number", "description": "Fee rate as decimal", "example": 0.001 },
              "symbol": { "type": "string", "example": "DUSD" }
            }
          }
        }
      },
      "DUSDRedeemRequest": {
        "type": "object",
        "required": ["sender", "recipientSol", "amount", "outputMint"],
        "properties": {
          "sender": {
            "type": "string",
            "minLength": 20,
            "maxLength": 64,
            "description": "DCC wallet address burning DUSD"
          },
          "recipientSol": {
            "type": "string",
            "minLength": 32,
            "maxLength": 44,
            "description": "Solana wallet address to receive USDC/USDT"
          },
          "amount": {
            "type": "number",
            "minimum": 1,
            "description": "Amount of DUSD to redeem"
          },
          "outputMint": {
            "type": "string",
            "enum": [
              "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
              "Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB"
            ],
            "description": "Output SPL token mint. Choose USDC or USDT."
          }
        }
      },
      "DUSDRedeemResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "redeem": {
            "type": "object",
            "properties": {
              "recipientSol": { "type": "string", "description": "Solana recipient address" },
              "outputMint": { "type": "string", "description": "Output token mint address" },
              "redeemAmount": { "type": "string", "description": "USDC/USDT to be unlocked (after fee)", "example": "99.900000" },
              "feeAmount": { "type": "string", "description": "Fee deducted", "example": "0.100000" },
              "feeRate": { "type": "number", "description": "Fee rate as decimal", "example": 0.001 },
              "symbol": { "type": "string", "example": "DUSD" }
            }
          }
        }
      },
      "DUSDInfoResponse": {
        "type": "object",
        "properties": {
          "symbol": { "type": "string", "example": "DUSD" },
          "name": { "type": "string", "example": "Decentral USD" },
          "fullName": { "type": "string", "example": "Decentral USD Stablecoin" },
          "decimals": { "type": "integer", "example": 6 },
          "description": { "type": "string" },
          "backing": {
            "type": "object",
            "properties": {
              "acceptedTokens": { "type": "array", "items": { "type": "string" }, "example": ["USDC", "USDT"] },
              "mechanism": { "type": "string", "example": "1:1 collateralized" },
              "vaultAddress": { "type": "string", "description": "Solana vault address holding DUSD reserves" }
            }
          },
          "fees": {
            "type": "object",
            "properties": {
              "mintFee": { "type": "string", "example": "0.10%" },
              "redeemFee": { "type": "string", "example": "0.10%" },
              "minimumMint": { "type": "number", "example": 1 },
              "minimumRedeem": { "type": "number", "example": 1 }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error message" },
          "details": {
            "type": "array",
            "description": "Validation error details (Zod issues)",
            "items": {
              "type": "object",
              "properties": {
                "code": { "type": "string" },
                "message": { "type": "string" },
                "path": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Too many requests — rate limit exceeded" }
          }
        }
      },
      "BridgePaused": {
        "description": "Bridge is currently paused",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Bridge is currently paused" }
          }
        }
      }
    }
  }
}